fix: move convertNetworkToNetworkCmdName
