+++ src/commands/config/create.ts
@@ -4,8 +4,9 @@ import { error, info, success } from "signale";
 import { Argv } from "yargs";
 import { create } from "../../implementations/config/create";
 import { getLogger } from "../../logger";
-import { convertNetworkToNetworkCmdName, highlight } from "../../utils";
+import { highlight } from "../../utils";
 import { CreateConfigCommand, TestNetwork } from "./config.type";
+import { NetworkCmdName } from "../networks";
 
 const { trace } = getLogger("config:create");
 
@@ -75,4 +76,14 @@ export const handler = async (args: CreateConfigCommand): Promise<void> => {
       error(e.message);
     }
   }
+};
+
+const convertNetworkToNetworkCmdName = (selectedNetwork: TestNetwork): NetworkCmdName => {
+  const network = {
+    [TestNetwork.Local]: NetworkCmdName.Local,
+    [TestNetwork.Goerli]: NetworkCmdName.Goerli,
+    [TestNetwork.Sepolia]: NetworkCmdName.Sepolia,
+    [TestNetwork.Mumbai]: NetworkCmdName.Maticmum,
+  };
+  return network[selectedNetwork];
 };
